[drc] -Fix: count view list hits in long so -1 can be returned
git-svn-id: svn://repo.hu/pcb-rnd/trunk@21276 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/view.c
@@ -81,10 +81,10 @@ pcb_view_t *pcb_view_by_uid(const pcb_view_list_t *lst, unsigned long int uid)
 	return NULL;
 }
 
-pcb_view_t *pcb_view_by_uid_cnt(const pcb_view_list_t *lst, unsigned long int uid, size_t *cnt)
+pcb_view_t *pcb_view_by_uid_cnt(const pcb_view_list_t *lst, unsigned long int uid, long *cnt)
 {
 	pcb_view_t *v;
-	size_t c = 0;
+	long c = 0;
 
 	for(v = pcb_view_list_first((pcb_view_list_t *)lst); v != NULL; v = pcb_view_list_next(v), cnt++) {
 		if (v->uid == uid) {
